chore: move Wildcard and NAIA into canonical package
