+++ easyuse_anima/naia/__init__.py
@@ -1,0 +1,3 @@
+"""NAIA feature package boundary."""
+
+__all__ = ()
