+++ tests/test_nodes_module_analyzer.py
@@ -73,15 +73,16 @@ def load_dynamic():
     def test_current_nodes_module_shape_matches_recorded_baseline(self):
         report = analyzer.analyze_path(ROOT / "nodes.py")
 
-        self.assertEqual(report["git_blob_sha1"], "8e19ca24696f778904942382d143415d033a2b82")
-        # Issue #184 B-04 moved the image/Detailer vertical slice as direct aliases.
-        self.assertEqual(report["top_level"]["function_count"], 260)
-        self.assertEqual(report["top_level"]["class_count"], 22)
-        self.assertEqual(report["line_count"], 12_131)
+        self.assertEqual(report["git_blob_sha1"], "b71aa17bd93a561b3597a7b9bd11296cb32dfaab")
+        # Issue #184 B-05 moved the Wildcard/NAIA vertical slice as direct aliases.
+        self.assertEqual(report["top_level"]["function_count"], 241)
+        self.assertEqual(report["top_level"]["class_count"], 20)
+        self.assertEqual(report["line_count"], 11_173)
         class_names = {item["name"] for item in report["top_level"]["classes"]}
         self.assertIn("EasyUseAnimaAIOGenerator", class_names)
         self.assertIn("EasyUseAnimaPromptStudioAdvancedV2", class_names)
-        self.assertIn("EasyUseAnimaNAIARandomPrompt", class_names)
+        self.assertNotIn("EasyUseAnimaNAIARandomPrompt", class_names)
+        self.assertNotIn("EasyUseAnimaWildcard", class_names)
         self.assertNotIn("EasyUseAnimaImageScaleByMultiple", class_names)
         self.assertNotIn("EasyUseAnimaDetailerAlignHook", class_names)
